+++ src/functions/emit/diff_generator.c
@@ -33,10 +33,10 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
     if (g_cdd_fail_alloc == 5555)
       diff_buf = NULL;
     else
-      diff_buf = (char *)malloc(diff_cap);
+      diff_buf = (char *)(size_t)malloc(diff_cap);
   }
 #else
-  diff_buf = (char *)malloc(diff_cap);
+  diff_buf = (char *)(size_t)malloc(diff_cap);
 #endif
   if (!diff_buf) {
     C_CDD_LOG_DEBUG("ENOMEM: OOM\n");
@@ -87,9 +87,9 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
             if (g_cdd_fail_alloc == 6666)
               new_buf = NULL;
             else
-              new_buf = (char *)realloc(diff_buf, diff_cap);
+              new_buf = (char *)(size_t)realloc(diff_buf, diff_cap);
 #else
-            char *new_buf = (char *)realloc(diff_buf, diff_cap);
+            char *new_buf = (char *)(size_t)realloc(diff_buf, diff_cap);
 #endif
             if (!new_buf) {
               free(diff_buf);
@@ -117,9 +117,9 @@ cdd_c_error_t patch_list_generate_diff(const struct TokenList *tokens,
         if (g_cdd_fail_alloc == 7777)
           new_buf = NULL;
         else
-          new_buf = (char *)realloc(diff_buf, diff_cap);
+          new_buf = (char *)(size_t)realloc(diff_buf, diff_cap);
 #else
-        char *new_buf = (char *)realloc(diff_buf, diff_cap);
+        char *new_buf = (char *)(size_t)realloc(diff_buf, diff_cap);
 #endif
         if (!new_buf) {
           free(diff_buf);
